fix: move litellm tests to integration, fix spacy install, drop -m slow
- Move 33 TestLiteLLMProvider + 3 factory skip tests to
  tests/integration/test_litellm_integration.py (require litellm installed)
- Fix spacy model install: use uv pip with direct wheel URL
  (spacy download requires pip which uv doesn't provide)
- Remove -m slow filter so all integration tests run
- Unit suite now has 2 skips (Windows platform + HF_TOKEN gated)
  down from 35
